move AxisControlWidget to bapsf_motion.gui.configure.controllers
